Add: Support AUTO_MODERATION_CONFIGURATION
