+++ lib/discorb.rb
@@ -49,7 +49,7 @@ require_order = %w[common flag dictionary error rate_limit http intents emoji_ta
                 %w[application audit_logs color components event event_handler automod] +
                 %w[attachment guild_template image integration interaction invite permission] +
                 %w[presence reaction role sticker utils voice_state webhook] +
-                %w[gateway_requests gateway app_command] +
+                %w[gateway_requests gateway_events gateway app_command] +
                 %w[asset extension shard client extend]
 require_order.each do |name|
   require_relative "discorb/#{name}.rb"
